Create new databases context file
