+++ lib/trento/discovery.ex
@@ -20,7 +20,7 @@ defmodule Trento.Discovery do
     SapSystemPolicy
   }
 
-  alias Trento.{Clusters, SapSystems}
+  alias Trento.{Clusters, Databases, SapSystems}
 
   @type command :: struct
 
@@ -142,7 +142,7 @@ defmodule Trento.Discovery do
 
   defp do_handle(%{"discovery_type" => "saptune_discovery", "agent_id" => agent_id} = event) do
     current_application_instances = SapSystems.get_application_instances_by_host_id(agent_id)
-    current_database_instances = SapSystems.get_database_instances_by_host_id(agent_id)
+    current_database_instances = Databases.get_database_instances_by_host_id(agent_id)
     sap_running = length(current_application_instances) + length(current_database_instances) > 0
 
     HostPolicy.handle(event, sap_running)
@@ -150,7 +150,7 @@ defmodule Trento.Discovery do
 
   defp do_handle(%{"discovery_type" => "sap_system_discovery", "agent_id" => agent_id} = event) do
     current_application_instances = SapSystems.get_application_instances_by_host_id(agent_id)
-    current_database_instances = SapSystems.get_database_instances_by_host_id(agent_id)
+    current_database_instances = Databases.get_database_instances_by_host_id(agent_id)
     cluster_id = Trento.Clusters.get_cluster_id_by_host_id(agent_id)
 
     SapSystemPolicy.handle(
