+++ lib/trento/sap_systems.ex
@@ -1,20 +1,16 @@
 defmodule Trento.SapSystems do
   @moduledoc """
-  Provides a set of functions to interact with SAP systems and HANA Databases.
+  Provides a set of functions to interact with SAP systems.
   """
 
   import Ecto.Query
 
   alias Trento.SapSystems.Projections.{
     ApplicationInstanceReadModel,
-    DatabaseInstanceReadModel,
-    DatabaseReadModel,
     SapSystemReadModel
   }
 
   alias Trento.Support.DateService
-
-  alias Trento.Databases.Commands.DeregisterDatabaseInstance
 
   alias Trento.SapSystems.Commands.DeregisterApplicationInstance
 
@@ -33,29 +29,10 @@ defmodule Trento.SapSystems do
     ])
   end
 
-  @spec get_all_databases :: [DatabaseReadModel.t()]
-  def get_all_databases do
-    DatabaseReadModel
-    |> where([d], is_nil(d.deregistered_at))
-    |> order_by(asc: :sid)
-    |> Repo.all()
-    |> Repo.preload([
-      :database_instances,
-      :tags
-    ])
-  end
-
   @spec get_application_instances_by_host_id(String.t()) :: [ApplicationInstanceReadModel.t()]
   def get_application_instances_by_host_id(host_id) do
     ApplicationInstanceReadModel
     |> where([a], a.host_id == ^host_id)
-    |> Repo.all()
-  end
-
-  @spec get_database_instances_by_host_id(String.t()) :: [DatabaseInstanceReadModel.t()]
-  def get_database_instances_by_host_id(host_id) do
-    DatabaseInstanceReadModel
-    |> where([d], d.host_id == ^host_id)
     |> Repo.all()
   end
 
@@ -79,34 +56,6 @@ defmodule Trento.SapSystems do
         commanded().dispatch(
           DeregisterApplicationInstance.new!(%{
             sap_system_id: sap_system_id,
-            host_id: host_id,
-            instance_number: instance_number,
-            deregistered_at: date_service.utc_now()
-          })
-        )
-    end
-  end
-
-  @spec deregister_database_instance(Ecto.UUID.t(), Ecto.UUID.t(), String.t(), DateService) ::
-          :ok | {:error, :instance_present} | {:error, :database_instance_not_registered}
-  def deregister_database_instance(
-        sap_system_id,
-        host_id,
-        instance_number,
-        date_service \\ DateService
-      ) do
-    case Repo.get_by(DatabaseInstanceReadModel,
-           sap_system_id: sap_system_id,
-           host_id: host_id,
-           instance_number: instance_number
-         ) do
-      %DatabaseInstanceReadModel{absent_at: nil} ->
-        {:error, :instance_present}
-
-      _ ->
-        commanded().dispatch(
-          DeregisterDatabaseInstance.new!(%{
-            database_id: sap_system_id,
             host_id: host_id,
             instance_number: instance_number,
             deregistered_at: date_service.utc_now()
